requirements.txt
asgiref==3.3.4
cloudinary==1.25.0
Django==3.2.3
django-cloudinary-storage==0.3.0
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
